Finishing todos

+++ src/order.java
@@ -14,14 +14,15 @@ public class Order {
     private final int orderID;       
     private int[] prodID;            
     private int[] amount;            
-    private int[] pricePerID;       
+    private double[] pricePerID;       
     private double totalPrice;       
     private String clientAddress;    
     private LocalDate orderDate;      
     private LocalDate arrivalDate;    
 
-    
-    public Order(int[] prodID, int[] amount, int[] pricePerID, String clientAddress, LocalDate orderDate, LocalDate arrivalDate) {
+
+
+    public Order(int[] prodID, int[] amount, double[] pricePerID, String clientAddress, LocalDate orderDate, LocalDate arrivalDate) {
         this.orderID = idCounter++;  
         this.prodID = prodID;
         this.amount = amount;
@@ -58,11 +59,11 @@ public class Order {
     }
 
     
-    public int[] getPricePerID() {
+    public double[] getPricePerID() {
         return pricePerID;
     }
 
-    public void setPricePerID(int[] pricePerID) {
+    public void setPricePerID(double[] pricePerID) {
         this.pricePerID = pricePerID;
         calculateTotalPrice();
     }
@@ -145,7 +146,7 @@ public class Order {
                         detailsStmt.setInt(1, orderID);
                         detailsStmt.setInt(2, prodID[i]);
                         detailsStmt.setInt(3, amount[i]);
-                        detailsStmt.setInt(4, pricePerID[i]);
+                        detailsStmt.setDouble(4, pricePerID[i]);
                         detailsStmt.addBatch(); // Batch processing for efficiency
                     }
                     detailsStmt.executeBatch();
